refactor: buildAndExecuteProposal into supports
